api/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==4.1.2
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycparser==2.21
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
